Accidents and Road Safety Analysis
Context.
This project is based on real world data; specifically, road traffic accidents in 2019.

Project Background Information.
All road traffic accidents involving casualties are logged and reported in the United Kingdom, along with (probably) a majority of other non-fatal road traffic accidents. Every year, the government releases a large batch of data associated with these reports. In this project I will be using the data
from 2019 as it represents a very complete sample with a lot of ancillary data available. The data contains the following files:
(1) Brief-guide-to road-accidents-and-safety-data. This file contains a very short introduction to the data set produced by the government.
(2) variable lookup. This spreadsheet details the large number of variables present in each of the datasets.
(3) Road Safety Data - Accidents 2019. This spreadsheet contains the details of the accidents logged in 2019.
(4) Road Safety Data - Casualties 2019. This spreadsheet details the casualties contained in the logged accidents.
(5) Road Safety Data- Vehicles 2019. This spreadsheet contains the details of the vehicles involved in the accidents.
(6) Adjustment Files. These files contain government modelling for the probabilities of different injuries occurring for each accident.
(i) adjustment figure guidance. This provides a short amount of government guidance on the figures included.
(ii) cas_adjustment_lookup_2019. This spreadsheet contains the tabulated adjusted probabilities of injuries.

Some of these files are simply too large to open even within excel. Therefore, it will be necessary to undertake analysis within Python.

The Task.
Advise government agencies about how to improve road safety and create a model that would predict such accidents and the injuries that they incur.